Refactor video/audit and move common functionality to handling audio/video stream information into separate video/media_info.py
